requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.3.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.4.26=hbd8a1cb_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
google-auth=2.40.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.71.0=pypi_0
idna=3.10=pypi_0
json5=0.9.6=pypi_0
ld_impl_linux-64=2.43=h712a8e2_4
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgomp=15.1.0=h767d61c_2
liblzma=5.8.1=hb9d3cd8_1
liblzma-devel=5.8.1=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libsqlite=3.50.0=hee588c1_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
markdown=3.8=pypi_0
markupsafe=3.0.2=pypi_0
ncurses=6.5=h2d0b736_3
numpy=1.26.0=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.5.0=h7b32b05_1
pillow=10.0.1=pypi_0
pip=25.1.1=pyh8b19718_0
protobuf=6.31.1=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
python=3.10.13=hd12c33a_1_cpython
pyyaml=6.0=pypi_0
readline=8.2=h8c095d6_2
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9.1=pypi_0
scipy=1.11.3=pypi_0
setuptools=80.8.0=pyhff2d567_0
six=1.17.0=pypi_0
tensorboard=2.14.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
timm=0.4.12=pypi_0
tk=8.6.13=noxft_hd72426e_102
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025b=h78e105d_0
urllib3=2.4.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
xz=5.8.1=hbcc6ac9_1
xz-gpl-tools=5.8.1=hbcc6ac9_1
xz-tools=5.8.1=hb9d3cd8_1
